Create CoreTerm class — move methods between Term and Menu
